# test-results/2025-09-03/crawler-tests/data/openapi_schema_170814.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "历史文本项目 - 数据源服务",
    "description": "\n    ## 数据源服务 API\n    \n    提供多平台内容爬取、代理管理和内容处理功能的微服务。\n    \n    ### 主要功能\n    - 🕷️ **爬虫管理**: 支持今日头条、百家号、小红书等平台\n    - 📄 **内容管理**: 手动添加、批量导入、查询和更新内容\n    - 🌐 **代理管理**: 代理池管理、自动测试和轮换\n    - 📊 **监控统计**: 实时状态监控和数据统计\n    \n    ### 技术栈\n    - **框架**: FastAPI + Python 3.9+\n    - **数据库**: MongoDB + Redis\n    - **爬虫**: Scrapy + Selenium\n    - **代理**: 多供应商代理池\n    \n    ### 使用说明\n    1. 使用 `/crawlers/` 接口管理爬虫任务\n    2. 使用 `/content/` 接口管理内容数据\n    3. 使用 `/proxy/` 接口管理代理设置\n    4. 查看 `/health` 检查服务状态\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/content/batch": {
      "post": {
        "tags": [
          "内容管理"
        ],
        "summary": "批量添加内容",
        "description": "批量添加多个内容\n\n- **contents**: 内容列表（最多100个）\n- **batch_name**: 批次名称（可选）\n- **auto_deduplicate**: 是否自动去重（默认为true）",
        "operationId": "create_batch_content_api_v1_content_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentBatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Batch Content Api V1 Content Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/upload": {
      "post": {
        "tags": [
          "内容管理"
        ],
        "summary": "文件导入内容",
        "description": "通过文件导入内容（支持CSV和JSON格式）\n\nCSV格式要求：\n- title: 标题（必填）\n- content: 内容（必填）  \n- source: 来源（必填）\n- author: 作者（可选）\n- source_url: 原始链接（可选）\n- keywords: 关键词（逗号分隔，可选）\n\nJSON格式要求：\n- 数组格式，每个对象包含上述字段",
        "operationId": "upload_content_file_api_v1_content_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_content_file_api_v1_content_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Upload Content File Api V1 Content Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{content_id}": {
      "get": {
        "tags": [
          "内容管理"
        ],
        "summary": "获取内容详情",
        "description": "获取指定内容的详细信息",
        "operationId": "get_content_detail_api_v1_content__content_id__get",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "内容ID",
              "title": "Content Id"
            },
            "description": "内容ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Content Detail Api V1 Content  Content Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "内容管理"
        ],
        "summary": "更新内容",
        "description": "更新指定内容的信息",
        "operationId": "update_content_api_v1_content__content_id__put",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "内容ID",
              "title": "Content Id"
            },
            "description": "内容ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Content Api V1 Content  Content Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "内容管理"
        ],
        "summary": "删除内容",
        "description": "删除指定的内容",
        "operationId": "delete_content_api_v1_content__content_id__delete",
        "parameters": [
          {
            "name": "content_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "内容ID",
              "title": "Content Id"
            },
            "description": "内容ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Content Api V1 Content  Content Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_content_file_api_v1_content_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "上传文件（支持CSV、JSON格式）"
          },
          "batch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Name",
            "description": "批次名称"
          },
          "auto_deduplicate": {
            "type": "boolean",
            "title": "Auto Deduplicate",
            "description": "是否自动去重",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_content_file_api_v1_content_upload_post"
      },
      "ContentBatchCreate": {
        "properties": {
          "contents": {
            "items": {
              "$ref": "#/components/schemas/ContentCreate"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Contents",
            "description": "内容列表"
          },
          "batch_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Name",
            "description": "批次名称"
          },
          "auto_deduplicate": {
            "type": "boolean",
            "title": "Auto Deduplicate",
            "description": "是否自动去重",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "contents"
        ],
        "title": "ContentBatchCreate",
        "description": "批量创建内容请求模型"
      },
...
      "ContentUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          }
        },
        "type": "object",
        "title": "ContentUpdate",
        "description": "更新内容请求模型"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}